requirements.txt
fastapi
uvicorn
supabase
pandas
scikit-learn
nltk
python-dotenv
rapidfuzz
sentence-transformers